requirements.txt
fastapi==0.111.0
langchain==0.2.0
langchain-chroma==0.1.1
langchain-community==0.2.0
langchain-core==0.2.0
langchain-google-genai==1.0.4
langchain-text-splitters==0.2.0
pydantic==2.7.1
python-docx==1.1.2
python-dotenv==1.0.1
sentence-transformers